Update README.md
The main aim of this project is to develop an engaging browser-based game that tests the player's cognitive speed and focus by challenging them to identify the color of the text, not the word itself. This helps users improve their attention, multitasking, and quick decision-making skills in a fun and interactive way using only HTML, CSS, and JavaScript.